defui init

+++ defui/src/modifiers.rs
@@ -5,7 +5,7 @@ use std::marker::PhantomData;
 /// Modifiers common to all views.
 pub trait Modifiers: View + Sized {
     /// Calls a closure after rendering with context and delta time.
-    fn anim<F: Fn(&mut Context, f32) + 'static + Clone>(self, func: F) -> AnimView<Self, F> {
+    fn anim<F: Fn(&mut Context, f64) + 'static + Clone>(self, func: F) -> AnimView<Self, F> {
         AnimView::new(self, func)
     }
